requirements.txt
# Core
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Environment
python-dotenv>=1.0.0

# Market data
yfinance>=0.2.28

# HTTP / feeds
requests>=2.31.0
feedparser>=6.0.10

# LLM providers
anthropic>=0.20.0
openai>=1.12.0

# Scheduling (local)
apscheduler>=3.10.0

# TradingAgents -- install from GitHub (not on PyPI)
# This also installs all its dependencies (langgraph, langchain-anthropic, etc.)
git+https://github.com/TauricResearch/TradingAgents.git
